+++ src/realm/util/client_websocket.cpp
@@ -1,3 +1,20 @@
+/*************************************************************************
+ *
+ * Copyright 2022 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ **************************************************************************/
 #include <realm/util/client_websocket.hpp>
 
 #include <realm/util/websocket.hpp>
@@ -9,7 +26,7 @@ namespace realm::util::websocket {
 namespace {
 class SocketImpl final : public WebSocket, public realm::util::websocket::Config {
 public:
-    SocketImpl(SocketConfig& config, SocketObserver& observer, Endpoint&& endpoint)
+    SocketImpl(SocketFactoryConfig& config, SocketObserver& observer, Endpoint&& endpoint)
         : m_config(config)
         , m_observer(observer)
         , m_endpoint(std::move(endpoint))
@@ -97,7 +114,7 @@ private:
         return m_config.logger;
     }
 
-    SocketConfig& m_config;
+    SocketFactoryConfig& m_config;
     SocketObserver& m_observer;
 
     const Endpoint m_endpoint;