+++ test/object-store/util/test_file.hpp
@@ -222,8 +222,7 @@ public:
     };
 
     TestSyncManager(realm::SyncManager::MetadataMode mode);
-    TestSyncManager(const Config& = Config(), const SyncServer::Config& = {});
-    TestSyncManager(const realm::SyncClientConfig&, const Config& = Config(), const SyncServer::Config& = {});
+    TestSyncManager(const Config& = Config(), const SyncServer::Config& = {}, const realm::util::Optional<realm::SyncClientConfig> = realm::util::none);
     ~TestSyncManager();
 
     std::shared_ptr<realm::app::App> app() const noexcept